Bug修復 1.管理員ID5為不可刪除的以及修改的資料，因為怕說有看專題的人真的會去刪除資料，如果真的刪除了，那麼看專題的人會沒有賬號密碼進入後台操作 2.新增Admin以及修改Admin密碼修改以及新增的驗證。新增未輸入內容的驗證 3.新增商品新增以及修改有空值無法進行修改機制。 4.新增USer管理，新增以及修改欄位有空值不能進行修改以及新增機制。

+++ TravelProject1.0/TravelProject1.0/Areas/Admin/Controllers/Api/SuperApiController.cs
@@ -64,6 +64,7 @@ public class SuperApiController : ControllerBase
     {
         var admin = await _context.Admins.FindAsync(id);
         if (admin == null) return false;
+        if (admin.Id == 5) return false;
         admin.Name = apDto.Name;
         admin.Account = apDto.Account;
         admin.Describe = apDto.Describe;
@@ -78,6 +79,7 @@ public class SuperApiController : ControllerBase
     {
         var admin = await _context.Admins.FindAsync(id);
         if (admin == null) return false;
+        if (admin.Id == 5) return false; 
         try
         {
             _context.Admins.Remove(admin);
